# GGRC-5848, GGRC-5851 Rewrite index records insert to row SQL