@olegnatsevsky 2018-10-12
reindex performance:
dev - 1421c
pr  - 866c